+++ Patches/GameLoadPatches.cs
@@ -73,6 +73,11 @@ namespace Localyssation.Patches
                 var key = KeyUtil.GetForAsset(skill);
                 Localyssation.defaultLanguage.RegisterKey($"{key}_NAME", skill._skillName);
                 Localyssation.defaultLanguage.RegisterKey($"{key}_DESCRIPTION", skill._skillDescription);
+                for (var rankIndex = 0; rankIndex < skill._skillRanks.Length; rankIndex++)
+                {
+                    var rank = skill._skillRanks[rankIndex];
+                    Localyssation.defaultLanguage.RegisterKey($"{key}_RANK_{rankIndex + 1}_DESCRIPTOR", rank._rankDescriptor);
+                }
             }
             foreach (var statAttribute in GameManager._current._statLogics._statAttributes)
             {
@@ -137,7 +142,7 @@ namespace Localyssation.Patches
 
             // scene-specific
             // this temporarily loads EVERY scene in the game to gather scene-specific keys, so we'll do it only when necessary
-            if (Localyssation.configCreateDefaultLanguageFiles.Value)
+            if (Localyssation.configTranslatorMode.Value && Localyssation.configCreateDefaultLanguageFiles.Value)
             {
                 Localyssation.instance.StartCoroutine(RegisterSceneSpecificStrings());
             }
@@ -145,7 +150,7 @@ namespace Localyssation.Patches
             // misc
             Localyssation.defaultLanguage.strings["FORMAT_QUEST_MENU_CELL_REWARD_CURRENCY"] = $"{{0}} {GameManager._current._statLogics._currencyName}";
 
-            if (Localyssation.configCreateDefaultLanguageFiles.Value)
+            if (Localyssation.configTranslatorMode.Value && Localyssation.configCreateDefaultLanguageFiles.Value)
                 Localyssation.defaultLanguage.WriteToFileSystem();
         }