+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.3
+* The mod now requires [EasySettings](https://thunderstore.io/c/atlyss/p/Nessie/EasySettings/)
+* Moved the Language selection button and "Developer" configs to **Settings > Mods > Localyssation**
+* "Developer" configs were renamed to "Translator" configs, and are now disabled by default. They can be re-enabled by checking "Translator Mode" in the mod settings
+* Added new buttons for translators:
+	* Add Missing Keys to Current Language - finds strings that exist in the default language, but not in the currently selected language, and saves them to the current language's file
+	* Log Untranslated Strings - finds every string that is the same in both the default and the selected languages, and prints them to the BepInEx log output
+* Skills are now translatable
+* Added `SKILL_<name>_RANK_<rank number>_DESCRIPTOR` strings for skills' descriptions on certain ranks
+* Added optional `OF` variant for player class strings
+	* This variant gets used in the Skills tab for the "<class name> Skillbook" and "<class name> Skills" strings
+* Fixed the Reset button on some settings not being translated
 ## 0.0.2
 * Fixed Nickname, Race, Class, Level and Experience strings not getting translated in the Stats tab menu
 * Fixed FORMAT_QUEST_PROGRESS_CREEPS_KILLED being forced to use the _VARIANT_MANY variant even if the current language doesn't have one